enable to implement custom parser
