+++ src/test/java/io/socket/parser/Helpers.java
@@ -14,18 +14,17 @@ import static org.junit.Assert.assertThat;
 @RunWith(JUnit4.class)
 public class Helpers {
 
-    private static Parser.Encoder encoder = new Parser.Encoder();
+    private static Parser.Encoder encoder = new IOParser.Encoder();
     private static Packet<String> errorPacket = new Packet<String>(Parser.ERROR, "parser error");
 
     public static void test(final Packet obj) {
         encoder.encode(obj, new Parser.Encoder.Callback() {
             @Override
             public void call(Object[] encodedPackets) {
-                Parser.Decoder decoder = new Parser.Decoder();
-                decoder.on(Parser.Decoder.EVENT_DECODED, new Emitter.Listener() {
+                Parser.Decoder decoder = new IOParser.Decoder();
+                decoder.onDecoded(new Parser.Decoder.Callback() {
                     @Override
-                    public void call(Object... args) {
-                        Packet packet = (Packet)args[0];
+                    public void call(Packet packet) {
                         assertPacket(packet, obj);
                     }
                 });
@@ -35,11 +34,10 @@ public class Helpers {
     }
 
     public static void testDecodeError(final String errorMessage) {
-        Parser.Decoder decoder = new Parser.Decoder();
-        decoder.on(Parser.Decoder.EVENT_DECODED, new Emitter.Listener() {
+        Parser.Decoder decoder = new IOParser.Decoder();
+        decoder.onDecoded(new IOParser.Decoder.Callback() {
             @Override
-            public void call(Object... args) {
-                Packet packet = (Packet)args[0];
+            public void call(Packet packet) {
                 assertPacket(errorPacket, packet);
             }
         });
@@ -52,11 +50,10 @@ public class Helpers {
         encoder.encode(obj, new Parser.Encoder.Callback() {
             @Override
             public void call(Object[] encodedPackets) {
-                Parser.Decoder decoder = new Parser.Decoder();
-                decoder.on(Parser.Decoder.EVENT_DECODED, new Emitter.Listener() {
+                Parser.Decoder decoder = new IOParser.Decoder();
+                decoder.onDecoded(new Parser.Decoder.Callback() {
                     @Override
-                    public void call(Object... args) {
-                        Packet packet = (Packet)args[0];
+                    public void call(Packet packet) {
                         obj.data = originalData;
                         obj.attachments = -1;
                         assertPacket(packet, obj);
