+++ src/test/java/io/socket/parser/ParserTest.java
@@ -9,7 +9,7 @@ import org.junit.runners.JUnit4;
 @RunWith(JUnit4.class)
 public class ParserTest {
 
-    private static Parser.Encoder encoder = new Parser.Encoder();
+    private static Parser.Encoder encoder = new IOParser.Encoder();
 
     @Test
     public void encodeConnection() {
